pyproject.toml
[tool.poetry]
name = "chai-chat"
version = "0.1.0"
description = "chai-chat inc."
authors = ["ivan Chizhov"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.6"
uvicorn = "^0.32.1"
aiogram = "^3.16.0"
python-dotenv = "^1.0.1"
authx = "^1.4.1"
requests = "^2.32.3"
aiosqlite = "^0.20.0"
sqlalchemy = "^2.0.36"
mistralai = "^1.2.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
package-mode = false